Simple_Todo_List_ID456/src/main/resources/static/script.py
package com.example.controller;

import org.springframework.stereotype.Controller;
import org.springframework.web.bind.annotation.*;
import com.example.model.Todo;
import com.example.service.TodoService;
import java.util.List;

@Controller
@RequestMapping("/api/todos")
public class TodoController {

    private final TodoService todoService;

    public TodoController(TodoService todoService) {
        this.todoService = todoService;
    }

    @GetMapping
    @ResponseBody
    public List<Todo> getAllTodos() {
        return todoService.getAllTodos();
    }

    @PostMapping
    @ResponseBody
    public Todo addTodo(@RequestBody Todo todo) {
        return todoService.addTodo(todo);
    }

    @DeleteMapping("/{id}")
    @ResponseBody
    public void removeTodo(@PathVariable Long id) {
        todoService.removeTodo(id);
    }

    @GetMapping("/")
    public String index() {
        return "index";
    }
}